pyproject.toml
# SOFC AI Benchmark — project metadata and dependencies
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "sofc-bench"
version = "0.1.0"
description = "SOFC AI Benchmark framework"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0",
    "pyyaml",
    "jinja2",
    "python-dotenv",
    "openpyxl",
]

[tool.setuptools.packages.find]
where = ["src"]
